requirements.txt
Django~=4.2.16
PyYAML~=6.0.2